// src/app/opmobilitybackend/models/get-storage-case-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { GetLocationResponseDto } from '../models/get-location-response-dto';
import { GetProviderResponse } from '../models/get-provider-response';
import { ProductResponse } from '../models/product-response';
import { UploadedFileResponseDto } from '../models/uploaded-file-response-dto';
export interface GetStorageCaseResponseDto {
  createdAt: string;
  emptyWeightKg: number;
  heightMm: number;
  id: number;
  images: Array<UploadedFileResponseDto>;
  lengthMm: number;
  location: GetLocationResponseDto;
  locationId: number;
  manufacturedOn: string;
  name: string;
  observations: string;
  product: ProductResponse;
  productId: number;
  provider: GetProviderResponse;
  providerId: number;
  ref: string;
  status: 'Available' | 'AtEvent' | 'Inspecting' | 'UnderRepair' | 'Retired';
  type: 'WOODEN_BOX' | 'PLASTIC_BOX' | 'METAL_BOX' | 'CARTON_BOX' | 'OTHER';
  updatedAt: string;
  widthMm: number;
}